+++ src/old_pages/_document.tsx
@@ -1,0 +1,25 @@
+import React from 'react'
+import Document, { Html, Head, Main, NextScript } from 'next/document'
+
+export default class MyDocument extends Document {
+  render(): JSX.Element {
+    return (
+      <Html lang="pt">
+        <Head>
+          <link rel="shortcut icon" href="../favicon.png" type="image/png" />
+
+          <link rel="preconnect" href="https://fonts.gstatic.com" />
+          <link
+            href="https://fonts.googleapis.com/css2?family=Montserrat:wght@600;700&family=Nunito&display=swap"
+            rel="stylesheet"
+          />
+          <meta charSet="utf-8" />
+        </Head>
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    )
+  }
+}
