+++ src/old_pages/challenges/[slug].tsx
