+++ src/old_pages/rankings/challenge/[slug].tsx
